refactor(java): port history-propagation to tutorials/workflow Spring Boot layout
