+++ tutorials/workflow/java/history-propagation/history-propagation.http
@@ -1,0 +1,16 @@
+@apphost=http://localhost:8080
+
+### Start the PatientIntake workflow
+POST {{ apphost }}/start
+Content-Type: application/json
+
+{
+  "patientId": "P-1042",
+  "name": "Jane Doe",
+  "condition": "bacterial sinusitis",
+  "medication": "amoxicillin",
+  "dosage": 500
+}
+
+### Get the workflow output
+GET {{ apphost }}/output
