+++ tutorials/workflow/java/history-propagation/src/main/java/io/dapr/springboot/examples/historypropagation/activities/CheckAllergiesActivity.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2026 The Dapr Authors
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.dapr.springboot.examples.historypropagation.activities;
+
+import io.dapr.springboot.examples.historypropagation.models.PatientRecord;
+import io.dapr.workflows.WorkflowActivity;
+import io.dapr.workflows.WorkflowActivityContext;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Component;
+
+@Component
+public class CheckAllergiesActivity implements WorkflowActivity {
+  private static final Logger logger = LoggerFactory.getLogger(CheckAllergiesActivity.class);
+
+  @Override
+  public Object run(WorkflowActivityContext ctx) {
+    PatientRecord rec = ctx.getInput(PatientRecord.class);
+    logger.info("Screening {} for {} allergies", rec.getPatientId(), rec.getMedication());
+    return true;
+  }
+}
