+++ tutorials/workflow/java/history-propagation/src/main/java/io/dapr/springboot/examples/historypropagation/models/DispenseResult.java
@@ -1,0 +1,54 @@
+/*
+ * Copyright 2026 The Dapr Authors
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.dapr.springboot.examples.historypropagation.models;
+
+public class DispenseResult {
+
+  private String dispenseId;
+  private String status;
+  private int reviewedWorkflows;
+
+  public DispenseResult() {
+  }
+
+  public DispenseResult(String dispenseId, String status, int reviewedWorkflows) {
+    this.dispenseId = dispenseId;
+    this.status = status;
+    this.reviewedWorkflows = reviewedWorkflows;
+  }
+
+  public String getDispenseId() {
+    return dispenseId;
+  }
+
+  public void setDispenseId(String dispenseId) {
+    this.dispenseId = dispenseId;
+  }
+
+  public String getStatus() {
+    return status;
+  }
+
+  public void setStatus(String status) {
+    this.status = status;
+  }
+
+  public int getReviewedWorkflows() {
+    return reviewedWorkflows;
+  }
+
+  public void setReviewedWorkflows(int reviewedWorkflows) {
+    this.reviewedWorkflows = reviewedWorkflows;
+  }
+}
