+++ tutorials/workflow/java/history-propagation/src/main/java/io/dapr/springboot/examples/historypropagation/models/InsuranceResult.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2026 The Dapr Authors
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.dapr.springboot.examples.historypropagation.models;
+
+public class InsuranceResult {
+
+  private boolean approved;
+  private String policyNumber;
+
+  public InsuranceResult() {
+  }
+
+  public InsuranceResult(boolean approved, String policyNumber) {
+    this.approved = approved;
+    this.policyNumber = policyNumber;
+  }
+
+  public boolean isApproved() {
+    return approved;
+  }
+
+  public void setApproved(boolean approved) {
+    this.approved = approved;
+  }
+
+  public String getPolicyNumber() {
+    return policyNumber;
+  }
+
+  public void setPolicyNumber(String policyNumber) {
+    this.policyNumber = policyNumber;
+  }
+}
