+++ tutorials/workflow/java/history-propagation/src/test/java/io/dapr/springboot/examples/TestHistoryPropagationApplication.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2026 The Dapr Authors
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.dapr.springboot.examples;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TestHistoryPropagationApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.from(HistoryPropagationApplication::main)
+            .with(DaprTestContainersConfig.class)
+            .run(args);
+    org.testcontainers.Testcontainers.exposeHostPorts(8080);
+  }
+}
