+++ tutorials/workflow/java/history-propagation/src/test/resources/application.properties
@@ -1,0 +1,1 @@
+spring.application.name=history-propagation
