--- workflows/java/sdk-context-propagation/dapr.yaml
@@ -1,7 +1,0 @@
-version: 1
-common:
-  resourcesPath: ../../components
-apps:
-  - appID: order-processor
-    appDirPath: ./order-processor/target
-    command: ["java", "-jar", "OrderProcessingContextPropagationService-0.0.1-SNAPSHOT.jar", "io.dapr.quickstarts.workflows.contextpropagation.WorkflowConsoleApp"]
