--- workflows/java/sdk-context-propagation/makefile
@@ -1,2 +1,0 @@
-include ../../../docker.mk
-include ../../../validate.mk
