--- workflows/java/sdk-context-propagation/order-processor/src/main/resources/simplelogger.properties
@@ -1,1 +1,0 @@
-org.slf4j.simpleLogger.logFile=System.out
